pyproject.toml
[project]
name = "saengsation"
version = "0.1.0"
description = "RGB LED controller for Keychron V7 — light sensation for your keyboard"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hidapi>=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
saengsation = "saengsation.cli:main"
